Bug 1526979 - Generate hg-push hooks. r=tomprince
This includes a hook task template, as well as a small Python script embedded
in a Docker image that creates decision tasks based on `.taskcluster.yml`

This is currently gated behind a temporary feature in projects.yml.  We can
slowly add this feature to various projects as we disable them in
mozilla-taskcluster, until everything is moved over.

Differential Revision: https://phabricator.services.mozilla.com/D21283