Bug 1534463: allow `vpn_hg_admin` group to set Mercurial related secrets r=tomprince
Now that we have given scopes to the correct contexts, we need
to provide the relevant parties with the ability to modify
the secrets. This commit grants write access for these secrets
to the `vpn_hg_admin` group. This group defines the set of
people with SSH access to the hgweb/hgssh cluster behind hgmo.

We also give this group the rights to change the `hgfingerprint`
secret. Previously this secret was managed by Greg, and I'm not sure
who else has access to change it. Now seems like an appropriate
time to give access to the relevant parties.

Differential Revision: https://phabricator.services.mozilla.com/D23981