Bug 1585644: revert idle timeout to 900s to see if we can get back closer to pre-migration levels. r=tomprince
After the TC migration, the utilization of workers got worse and we saw an increase in workers taking zero jobs before shutting down. I suspect something in the new AWS provider behaves differently than the old provisioner.

I want to revert the idle timeout back to where they were before the migration to see if we can improve overhead with the current provider, and to collect better metrics now.

Differential Revision: https://phabricator.services.mozilla.com/D60179