Bug 1611780: Remove duplication in listing available actions; r=mtabara
Originally, there was an entry in actions.yml for each trust_domain, level,
action_perm combination. However, for any trust domain, the all actions that
are present are available at every level, and the format makes it hard to
easily see what action there are for a given domain. Change this so that all
the actions for a given domain are in a single stanza

Differential Revision: https://phabricator.services.mozilla.com/D61111