Bug 1636297: Reduce the size of gecko-t/t-linux-* and related pools; r=jmaher
Reduce the size of linux tester pools to more closely match the historical
usage of the pools over the past 90 days.

For gecko-t/t-linux-large, we had a single peak of 2563 concurrent tasks once in the past 90d,
but are typically below 2000 concurrent tasks.

For gecko-t/t-linux-xlarge, we had peaks of 1667 and 1574 concurrent tasks, but
are typically below 1000 concurrent tasks.

For gecko-t/t-linux-xlarge-soruce, we had a peak of 200 councurrent tasks, but
are typically under 200 concurrent tasks.

Differential Revision: https://phabricator.services.mozilla.com/D80134