Bug 1639873: Add action permission to allow backfills to trigger actions; r=aki
Since we look at all action on a push to find tasks associated with that push, we
want to split the backfill action in two, with the action that schedules jobs on
a given push being an action on that push. This adds a new action permission to
allow the first action to trigger the hook for the second action.

Differential Revision: https://phabricator.services.mozilla.com/D80934