Bug 1664304 - Grant try access to civet and for Updatebot to retrigger tasks r=aki
While Updatebot will eventually run in try for testing; we will restrict
retrigger access to central (for production) and holly (for testing.)
On try, retriggers will fail, but we can work aorund that.

Civet can now run in try as coverity does; and like coverity it will
not be able to be run as an interactive task due to the scope whitelist.

Differential Revision: https://phabricator.services.mozilla.com/D90291