Bug 1741274 - stop setting fxci.utils.cli args.command to a string. r=releng-reviewers,gbrown
It looks like we have been adding a subparser with a dest of `command`, then setting the default `command` to a function, then calling the `command`. In earlier versions of python / pinned requirements, we got away with doing this. For some reason, the latest cloudops' docker image with our latest changes sets `command` to the string, then setting the default of `command` to the function pointer doesn't work, and we try to call the string with the arguments.

Let's stop setting the dest to `command`, and instead set it to `command_name`. We don't seem to refer to this anywhere, so we could leave it off and drop this information into the void [1]. In case we do care about it, let's leave it at `command_name`

[1] https://docs.python.org/3/library/argparse.html#sub-commands

Differential Revision: https://phabricator.services.mozilla.com/D131190