no bug - fix unpinned bustages. r=releng-reviewers,gabriel
When we call `pip install --user <package>`, we automatically pick up any new requirements for that package that aren't installed in our virtualenv, so let's use `--no-deps` so we always die if we don't have all of our dependencies installed with the initial requirements install.

Also, we didn't install taskgraph's pinned requirements in .tc.yml, so let's do that first.

Differential Revision: https://phabricator.services.mozilla.com/D142500