Bug 1775106 - Experiment with 'scalingRatio=0.5' for t-linux-metal, r=releng-reviewers,gbrown
This pool uses super beefy instances (e.g 96 cores) that are supposed to run up
to 32 tasks at a time. It is also used by the Android emulators which have a
very high setup / teardown cost.

For both these reasons, I think it would be more efficient if we try to reduce
the total number of instances we spawn. Turn down the scalingRatio to
accomplish this at the expense of possible longer pending times for tasks that
use this pool.

I propose we try this out for one month, and then evaluate the cost differences
of the pool and evalute whether we want to bump scalingRatio back up, keep it
the same, or reduce it even further.

Differential Revision: https://phabricator.services.mozilla.com/D150479