Bug 1784232 - part 3: Fix missing revision in cron jobs r=jcristau
I introduced this regression in 9616a7e794384173c6c4a2d6fbc5b568bf64815d - more precisely at this line[1]. Before then, `revision` was always defined because it was overridden by the value gotten from `json-pushes`. At that time, I defined `tip_revision` and thanks to the check above in the function, we're sure this value is identical to `revision` if `revision` is provided. `revision` is `None` in the case of a cron job. That's why it busted them all in bug 1788201.

[1] https://hg.mozilla.org/ci/ci-configuration/rev/9616a7e794384173c6c4a2d6fbc5b568bf64815d#l1.43

Differential Revision: https://phabricator.services.mozilla.com/D156080