Bug 1826577 - Convert 'mozilla-taskgraph' to a level 3 repo, r=releng-reviewers,jcristau
Initially I thought there was no need for this to be level 3 as we wouldn't
have release automation anyway, but there are some assumptions in Taskgraph
which break (around cached_tasks) when setting it up this way.

Ideally we should fix those assumptions, but for now making this level 3
unblocks things.

Differential Revision: https://phabricator.services.mozilla.com/D175068