Bug 1830149 - Remove app-services trust_project and add trust_domain_scopes feature, r=jcristau
This will grant some standard scopes that app-services was previously missing.
This feature grants a different set of scopes when 'trust_project' is present.
Typically 'trust_project' is only relevant when there are multiple projects in
a single trust_domain, so in this case removing it makes more sense to avoid
the unnecessarily verbose index routes.

Differential Revision: https://phabricator.services.mozilla.com/D176677