Bug 1833243 - [test] Add a fixture to lazily generate specific resources, r=releng-reviewers,bhearsum
This creates a new `generate_resources` fixture that can be used to generate
all, or only some resources. It can be used like this:

```python
async def test_foo(generate_resources):
    # generate resources from hooks.py, in_tree_actions.py, and cron.py only
    resources = await generate_resources("hooks", "in_tree_actions", "cron")

    # generate all resources
    resources = await generate_resources()
```

This fixture caches the resources from individual modules as it goes. So in the
above example, the resources from `hooks`, `in_tree_actions` and `cron` do not
get re-generated in the second call. Since the fixture is session-scoped, the
resources are also cached across all test modules and functions.

The `generated` fixture has been turned into an alias for `await
generate_resources()` so existing tests that use it can share caches with tests
generating only a subset of resources.

Differential Revision: https://phabricator.services.mozilla.com/D178110