No Bug - Stop granting shipit_firefox / app-services developers ability to create level 3 tasks, r=jcristau
There should be no reason to create L3 tasks directly, and allowing it could
even be a security concern.

Creating L1 tasks is likely not necessary, but could be useful if they want to
edit an existing task with a new environment variable or similar, so I don't
see the harm in leaving it.

Differential Revision: https://phabricator.services.mozilla.com/D179362