Bug 1906337 - Convert experimental dagger pool to generic-worker, r=releng-reviewers,bhearsum
After playing around with the `docker-worker` pool, I realized that the image I
was attempting to use was meant for a different purpose than what I thought it
was.

I looked at how they implemented the Github actions integration, and it seems
like it'll be best to run `dagger-engine` directly in the VM rather than in an
intermediary container.

Differential Revision: https://phabricator.services.mozilla.com/D215961