Bug 1903776: refactor projects.yml into new format that supports multiple branches r=releng-reviewers,ahal
I did my best to list out all of the used branches (aka, branches that need scopes) for each project explicitly in `branches`. Many projects _also_ contain an entry for a `*` branch for the moment to retain `*` or `branch:*` scopes for many roles. This is mostly to minimize risk when this lands. I would like to follow-up quickly to remove these `*` branches at least on GitHub, and thus remove the `*` and/or `branch:*` roles (otherwise we can't have different levels by branch).

With new branches being listed out explicitly we see a number of scope additions. As far as I can tell, this is not actually increasing access to anything, just granting these branches scopes that they already have through one of the `*` roles.

(My original aim here was to have absolutely _no_ change shown by `ci-admin diff`, but after a lot of fussing I don't think this is practical, so I've done my best to minimize it instead.)

Differential Revision: https://phabricator.services.mozilla.com/D214339