Bug 1907109: enable us-east-1 for translations v100 preemptible worker pools r=releng-reviewers,hneiva
This is the only other region in North America with v100s. We don't have quota issues with non-preemptible pools so far, so I'm avoiding adding this region to those, as us-east-1 is not a low CO2 region.

We have a quota limit of 32 for these right now. It can plausibly be raised to 128 if Google approves the same limit we got in the existing regions.

I'm also bumping maxCapacity to 96. Our quota is 128 in the existing 2 regions, and suspect it will get there for the new one as well. 128 / 4 = 32, 32*3 regions = 96 if we're only running 4 GPU instances.

Differential Revision: https://phabricator.services.mozilla.com/D216173